Renamed the folder Particles -> ParticleEmitters.
It doesn't contain Particles.h anymore and hasn't for a while now, so it
makes little sense to call the folder that.  It was renamed to reflect
its contents.

Note: Technically I made a new folder of that name and used "git mv" to
"rename" the files to their new folder.