After some tinkering, I found the problem.  All particles now stop on hitting a polygon edge.
The problem was my calculation of the "c" vector.  Given a vector <a1 +
a> and another <b1 + b> (a1 and b1 are points and "a" and "b" are
vectors from those points), then part of the "line segments intersect"
calculation involves the creation of a new directional (that is, not
positional) vector "c" = a1 - b1.  For the particle-geometry shader,
particle travel is "a" and the geometry line is "b".  I was calculating
"c" = b1 - a1 (vec4 c = polyFace._start._pos - p._prevPos).  Whoops.
Flipped that around, and now it works.  Yay!  Mystery solved.  I still
don't know how the math works though.