Fixed the heap error.
Somehow the linker set BvhNode::BoundingBox mixed up between the one in
BvhNode.h and the one in /Include/Geometry/BoundingBox.h.  It seemed
like space was used as if it were the geometry version, but space was
recorded as if it were the simple one in BvhNode.h.  Whatever happened,
I renamed the geometry version to Box2D and everything seems fine now.