Renamed a crap-ton of files.
The folders called ParticleCollisions for SSBOs and shaders were changed
to ParticleParticleCollisions.  The folders called
ParticleGeometryCollisions for SSBOs and shaders were changed to
ParticlePolygonCollisions.  All files with "CollidableGeometry" in the
name were renamed to "CollidablePolygon".  Now I need to fix the Visual
Studio project and fix the include paths and change all the buffer
names.