Collidable polygon BVH is now being constructed.
Found an error: The CollidablePolygonSsbo was not double sized, so the
"second half" that the sorting was trying to work with wasn't there, and
so SortCollidablePolygons.comp ended up dumping a bunch of zeros into
the first half.  I noticed the problem when my polygon BVH node bounding
boxes were all zeros.  Now it is fixed.

And now I need to clean up a bunch of debugging checks that I made.