Needed to create a "potential collisions" buffer for the collidable polygons as well.
I have have been able to reuse the
PotentialParticleParticleCollisionsBuffer in the
ParticlePolygonCollisions shader controller, but that would have
required me to either (1) create the SSBO outside the
ParticleParticleCollisions and ParticlePolygonCollisions shader
controllers and then pass it in or (2) have a public getter in
ParticleParticleCollisions and then have ParticlePolygonCollisions use
the getter.

I didn't like either option, so I made another SSBO (structure is
identical, only difference is that it is a different chunk of GPU memory
with a different binding) and stuck it into ParticlePolygonCollisions.