Got particle-polygon collisions detection working properly (I think).
The problem was in ParticlePolygonCollisions in which the uniform
uCollidablePolygonBvhNumberLeaves didn't get set for the
particle-polygon collision detection program, resulting in the shader
using the uniform default of 0 for the root polygon BVH node.  As a
result, the starting currentPolygonNodeIndex was 0, a leaf node, and
leaf nodes don't have any children, so the collidable polygons' BVH
wasn't being traversed.  That's a lot that can go wrong with a single
line of code :), but that collision detection is working now.  Now I
need to clean up after debugging and check that the collision resolution
is working properly.