Fixed it.
I was looping through all the collidable polygons while using
AllPotentialParticlePolygonCollisions[...]._numPotentialCollisions as
the loop limiter.

I am supposed to use that as the loop limiter, but then take the loop
counter as an index into the
AllPotentialParticlePolygonCollisions[..]._objectIndexes[...] array, and
then use THAT as the index into the collidable polygon array.  Whoops.

I had an epiphany when I hard-coded a check against polygon 0 and got
the same result.